+++ test/unit/mongo_class/code_coverage.sh
@@ -54,6 +54,7 @@ coverage run -a --source=mongo_class test/unit/mongo_class/server_is_primary.py
 coverage run -a --source=mongo_class test/unit/mongo_class/server_lock_db.py
 coverage run -a --source=mongo_class test/unit/mongo_class/server_set_pass_config.py
 coverage run -a --source=mongo_class test/unit/mongo_class/server_set_ssl_config.py
+coverage run -a --source=mongo_class test/unit/mongo_class/server_set_tls_config.py
 coverage run -a --source=mongo_class test/unit/mongo_class/server_unlock_db.py
 coverage run -a --source=mongo_class test/unit/mongo_class/server_upd_server_attr.py
 coverage run -a --source=mongo_class test/unit/mongo_class/server_upd_srv_stat.py