+++ test/unit/mongo_libs/code_coverage.sh
@@ -10,6 +10,7 @@ echo ""
 
 echo "Running unit test modules in conjunction with coverage"
 coverage run -a --source=mongo_libs test/unit/mongo_libs/add_ssl_cmd.py
+coverage run -a --source=mongo_libs test/unit/mongo_libs/add_tls_cmd.py
 coverage run -a --source=mongo_libs test/unit/mongo_libs/create_cmd.py
 coverage run -a --source=mongo_libs test/unit/mongo_libs/create_instance.py
 coverage run -a --source=mongo_libs test/unit/mongo_libs/crt_base_cmd.py