+++ test/unit/mongo_libs/unit_test_run3.sh
@@ -6,6 +6,7 @@
 
 echo "Unit test: mongo_class"  
 /usr/bin/python3 ./test/unit/mongo_libs/add_ssl_cmd.py
+/usr/bin/python3 ./test/unit/mongo_libs/add_tls_cmd.py
 /usr/bin/python3 ./test/unit/mongo_libs/create_cmd.py
 /usr/bin/python3 ./test/unit/mongo_libs/create_instance.py
 /usr/bin/python3 ./test/unit/mongo_libs/crt_base_cmd.py